Finalize main project from module 4

+++ modulo-04-dom/exercicios/principal/01-crud/main.js
@@ -1,4 +1,7 @@
-const baseUrl = "https://crudcrud.com/api/604dff26cc034dffb8adf6c0f00d059b";
+const baseUrl = "https://crudcrud.com/api/08bd560d12e64726845311cd46c3ec1c";
+
+const createForm = document.querySelector("form#create-form");
+const updateForm = document.querySelector("form#update-form");
 
 const petSpecies = {
   dog: "Cachorro",
@@ -37,6 +40,13 @@ const createPet = async (pet) => {
     body: JSON.stringify(pet),
   });
 
+  const pets = localStorage.getItem("m4-01-crud:pets");
+  let petsList = pets ? JSON.parse(pets) : [];
+
+  petsList.push(pet);
+
+  localStorage.setItem("m4-01-crud:pets", JSON.stringify(petsList));
+
   return response.json();
 };
 
@@ -49,6 +59,15 @@ const updatePet = async (pet, id) => {
     body: JSON.stringify(pet),
   });
 
+  const pets = localStorage.getItem("m4-01-crud:pets");
+  let petsList = pets ? JSON.parse(pets) : [];
+
+  const petIndex = petsList.findIndex((pet) => pet._id === id);
+
+  petsList[petIndex] = pet;
+
+  localStorage.setItem("m4-01-crud:pets", JSON.stringify(petsList));
+
   return response.json();
 };
 
@@ -56,11 +75,25 @@ const deletePet = async (id) => {
   await fetch(`${baseUrl}/pets/${id}`, {
     method: "DELETE",
   });
+
+  const pets = localStorage.getItem("m4-01-crud:pets");
+  let petsList = pets ? JSON.parse(pets) : [];
+
+  const petIndex = petsList.findIndex((pet) => pet._id === id);
+
+  petsList.splice(petIndex, 1);
 };
 
 const getPets = async () => {
   const response = await fetch(`${baseUrl}/pets`);
   const data = await response.json();
+
+  const pets = localStorage.getItem("m4-01-crud:pets");
+  let petsList = pets ? JSON.parse(pets) : [];
+
+  petsList = [...data];
+
+  localStorage.setItem("m4-01-crud:pets", JSON.stringify(petsList));
 
   return data;
 };
@@ -72,16 +105,115 @@ const getPet = async (id) => {
   return data;
 };
 
-const createUpdatePetModal = (pet) => {
-  
-}
+const renderUpdateFormModal = (pet) => {
+  const updateFormContainer = updateForm.parentElement;
+  updateFormContainer.classList.remove("hidden");
 
-(async () => {
+  const inputName = updateForm.querySelector("input[name=name]");
+  const inputBirthDate = updateForm.querySelector("input[name=birthDate]");
+  const selectKind = updateForm.querySelector("select[name=kind]");
+  const selectSize = updateForm.querySelector("select[name=size]");
+  const inputSex = updateForm.querySelector(`input[value=${pet.sex}]`);
+  const inputResponsibleCpf = updateForm.querySelector(
+    "input[name=responsibleCpf]"
+  );
+  const inputResponsiblePhone = updateForm.querySelector(
+    "input[name=responsiblePhone]"
+  );
+
+  inputName.value = pet.name;
+  inputBirthDate.value = pet.birthDate;
+  selectKind.value = pet.kind;
+  selectSize.value = pet.size;
+  inputSex.checked = true;
+  inputResponsibleCpf.value = pet.responsibleCpf;
+  inputResponsiblePhone.value = pet.responsiblePhone;
+
+  [
+    inputName,
+    inputBirthDate,
+    selectKind,
+    selectSize,
+    inputResponsibleCpf,
+    inputResponsiblePhone,
+  ].forEach((el) => {
+    el.classList =
+      "px-4 py-2 border border-gray-300 rounded-md outline-none focus:ring-2 focus:ring-gray-800 hover:ring-2 hover:ring-gray-800 transition-all ease-in duration-300";
+
+    const parent = el.parentElement;
+    parent.className = "flex flex-col";
+
+    const label = parent.querySelector("label");
+
+    const required = document.createElement("i");
+    required.innerText = "*";
+    required.classList = "text-red-500 text-sm";
+
+    label.appendChild(required);
+    label.classList = "text-gray-950 font-semibold";
+  });
+
+  const inputSexParent = inputSex.parentElement;
+  const inputSexParentContainer = inputSexParent.parentElement;
+  inputSexParentContainer.className = "flex flex-col gap-2";
+
+  const formattedDate = getNow().toISOString().split("T")[0];
+  inputBirthDate.setAttribute("max", formattedDate.split("T")[0]);
+
+  const inputRadios = updateFormContainer.querySelectorAll("input[type=radio]");
+  inputRadios.forEach((radio) => {
+    radio.classList = "mr-2 checked:accent-gray-900";
+  });
+
+  const updateButton = updateFormContainer.querySelector("button");
+  updateButton.classList =
+    "px-4 py-2 bg-gray-800 text-white font-semibold rounded-md hover:bg-gray-900 transition-all ease-in duration-300";
+
+  updateForm.addEventListener("submit", (event) => {
+    event.preventDefault();
+
+    const formData = new FormData(event.target);
+    const data = Object.fromEntries(formData);
+
+    Object.values(data).forEach((value) => {
+      if (value === "") {
+        alert("Todos os campos são obrigatórios");
+        return;
+      }
+    });
+
+    const formSize = Object.values(data).length;
+
+    if (formSize < 7) {
+      alert("Todos os campos são obrigatórios");
+      return;
+    }
+
+    if (!BrazilianValues.isCPF(data.responsibleCpf)) {
+      alert("CPF inválido");
+      return;
+    }
+
+    if (!BrazilianValues.isPhone(data.responsiblePhone)) {
+      alert("Telefone inválido");
+      return;
+    }
+
+    // updatePet(data, pet._id);
+
+    // location.reload();
+  });
+};
+
+const renderPets = async () => {
   const pets = await getPets();
 
   const table = document.querySelector("table#pets-table");
   const thead = table.querySelector("thead");
   const tbody = table.querySelector("tbody");
+
+  thead.innerHTML = null;
+  tbody.innerHTML = null;
 
   const trHead = document.createElement("tr");
   const thName = document.createElement("th");
@@ -90,7 +222,7 @@ const createUpdatePetModal = (pet) => {
   const thSize = document.createElement("th");
   const thSex = document.createElement("th");
   const thResponsiblePhone = document.createElement("th");
-  const thResponsibleEmail = document.createElement("th");
+  const thResponsibleCpf = document.createElement("th");
   const thActions = document.createElement("th");
 
   table.classList = "w-full table-fixed border-collapse mt-4";
@@ -101,7 +233,7 @@ const createUpdatePetModal = (pet) => {
   thSize.innerText = "Porte";
   thSex.innerText = "Sexo";
   thResponsiblePhone.innerText = "Telefone";
-  thResponsibleEmail.innerText = "Email";
+  thResponsibleCpf.innerText = "Cpf";
   thActions.innerText = "Ações";
 
   thName.classList = "text-start text-sm font-medium text-gray-900 pb-2";
@@ -110,12 +242,10 @@ const createUpdatePetModal = (pet) => {
   thSize.classList = "text-start text-sm font-medium text-gray-900";
   thSex.classList = "text-start text-sm font-medium text-gray-900";
   thResponsiblePhone.classList = "text-start text-sm font-medium text-gray-900";
-  thResponsibleEmail.classList = "text-start text-sm font-medium text-gray-900";
+  thResponsibleCpf.classList = "text-start text-sm font-medium text-gray-900";
   thActions.classList = "text-start text-sm font-medium text-gray-900";
 
   trHead.className = "border-b border-b-slate-800";
-
-  thResponsibleEmail.setAttribute("colspan", 2);
 
   trHead.appendChild(thName);
   trHead.appendChild(thBirthDate);
@@ -123,7 +253,7 @@ const createUpdatePetModal = (pet) => {
   trHead.appendChild(thSize);
   trHead.appendChild(thSex);
   trHead.appendChild(thResponsiblePhone);
-  trHead.appendChild(thResponsibleEmail);
+  trHead.appendChild(thResponsibleCpf);
   trHead.appendChild(thActions);
 
   thead.appendChild(trHead);
@@ -137,7 +267,7 @@ const createUpdatePetModal = (pet) => {
     const tdSize = document.createElement("td");
     const tdSex = document.createElement("td");
     const tdResponsiblePhone = document.createElement("td");
-    const tdResponsibleEmail = document.createElement("td");
+    const tdResponsibleCpf = document.createElement("td");
     const tdActions = document.createElement("td");
 
     const buttonEdit = document.createElement("button");
@@ -153,12 +283,16 @@ const createUpdatePetModal = (pet) => {
 
     buttonEdit.addEventListener("click", async () => {
       const petToUpdate = await getPet(pet._id);
-      createUpdatePetModal(petToUpdate);
+      renderUpdateFormModal(petToUpdate);
     });
 
     buttonDelete.addEventListener("click", async () => {
-      await deletePet(pet._id);
-      location.reload();
+      const shouldDelete = confirm("Deseja realmente excluir?");
+
+      if (shouldDelete) {
+        await deletePet(pet._id);
+        location.reload();
+      }
     });
 
     tdName.innerText = pet.name;
@@ -167,7 +301,7 @@ const createUpdatePetModal = (pet) => {
     tdSize.innerText = petSizes[pet.size];
     tdSex.innerText = petSex[pet.sex];
     tdResponsiblePhone.innerText = pet.responsiblePhone;
-    tdResponsibleEmail.innerText = pet.responsibleEmail;
+    tdResponsibleCpf.innerText = pet.responsibleCpf;
 
     const actionsContainer = document.createElement("div");
 
@@ -178,23 +312,26 @@ const createUpdatePetModal = (pet) => {
 
     tdActions.appendChild(actionsContainer);
 
-    tdResponsibleEmail.setAttribute("colspan", 2);
-
     tr.appendChild(tdName);
     tr.appendChild(tdBirthDate);
     tr.appendChild(tdType);
     tr.appendChild(tdSize);
     tr.appendChild(tdSex);
     tr.appendChild(tdResponsiblePhone);
-    tr.appendChild(tdResponsibleEmail);
+    tr.appendChild(tdResponsibleCpf);
     tr.appendChild(tdActions);
 
-    table.appendChild(tr);
+    tbody.appendChild(tr);
   }
+};
+
+(async () => {
+  await renderPets();
 })();
 
-const createForm = document.querySelector("form#create-form");
-const updateForm = document.querySelector("form#update-form");
+setInterval(async () => {
+  await renderPets();
+}, 5000);
 
 createForm.className =
   "flex flex-col gap-4 p-4 bg-gray-200 rounded-md w-full mt-10";
@@ -210,7 +347,7 @@ for (const el of createForm.children) {
     const inputText = el.querySelector("input[type=text]");
     const inputDate = el.querySelector("input[type=date]");
     const inputPhone = el.querySelector("input[type=number]");
-    const inputEmail = el.querySelector("input[type=email]");
+    const inputCpf = el.querySelector("input[type=email]");
     const inputRadio = el.querySelectorAll("input[type=radio]");
 
     const select = el.querySelector("select");
@@ -229,7 +366,7 @@ for (const el of createForm.children) {
     label.classList = "text-gray-950 font-semibold";
 
     inputText ? (inputText.classList = defaultInputStyle) : null;
-    inputEmail ? (inputEmail.classList = defaultInputStyle) : null;
+    inputCpf ? (inputCpf.classList = defaultInputStyle) : null;
     inputPhone ? (inputPhone.classList = defaultInputStyle) : null;
 
     if (inputDate) {
@@ -254,24 +391,36 @@ for (const el of createForm.children) {
 }
 
 createForm.addEventListener("submit", async (event) => {
-  console.log("aqui");
   event.preventDefault();
 
   const formData = new FormData(event.target);
   const data = Object.fromEntries(formData);
 
+  const formSize = Object.keys(data).length;
+
+  Object.values(data).forEach((value) => {
+    if (value === "") {
+      alert("Todos os campos são obrigatórios");
+      return;
+    }
+  });
+
+  if (formSize < 7) {
+    alert("Todos os campos são obrigatórios");
+    return;
+  }
+
+  if (!BrazilianValues.isCPF(data.responsibleCpf)) {
+    alert("CPF inválido");
+    return;
+  }
+
+  if (!BrazilianValues.isPhone(data.responsiblePhone)) {
+    alert("Telefone inválido");
+    return;
+  }
+
   await createPet(data);
 
-  createForm.reset();
+  location.reload();
 });
-
-// updateForm.addEventListener("submit", (event) => {
-//   event.preventDefault();
-
-//   const formData = new FormData(event.target);
-//   const data = Object.fromEntries(formData);
-
-//   console.log(data);
-
-//   createForm.reset();
-// });
